Some more elements
Time to copy paste from the thing:
"I had a backlog of elements to do, and I tried to do a few of them today. These elements are pretty complicated, but I was able to do a few of them."

"I’m okay with the work I did today, but I kinda got sidetracked because my head just hurts. I want to go to the neurologist to just see what’s going on, but it’s kind of delayed because of stuff, but I’m fine with what I did today."